requirement.txt
asgiref==3.9.1
Django==5.2.5
djangoql==0.18.1
djangorestframework==3.16.1
mysqlclient==2.2.7
ply==3.11
python-decouple==3.8
sqlparse==0.5.3
